refactor(error-handling): update CartService and PlatziStoreClient to use new exception responses
