+++ src/main/java/br/com/expresscart/client/PlatziStoreClient.java
@@ -1,7 +1,7 @@
 package br.com.expresscart.client;
 
 import br.com.expresscart.client.response.PlatziProductResponse;
-import br.com.expresscart.exceptions.CustomErrorDecoder;
+import br.com.expresscart.config.CustomErrorDecoder;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
