+++ src/main/java/br/com/expresscart/service/CartService.java
@@ -6,8 +6,8 @@ import br.com.expresscart.controller.request.PaymentRequest;
 import br.com.expresscart.entity.Cart;
 import br.com.expresscart.entity.Product;
 import br.com.expresscart.entity.Status;
-import br.com.expresscart.exceptions.BusinessException;
-import br.com.expresscart.exceptions.DataNotFoundException;
+import br.com.expresscart.exception.handler.BusinessException;
+import br.com.expresscart.exception.handler.DataNotFoundException;
 import br.com.expresscart.repository.CartRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
